Some bugs fixed
git-svn-id: file:///home/byron/Desktop/svn/archival/RealEstateMarketAnalysis/branches/version-0_75@185 6aeec067-7811-e111-a833-485b39788f86

+++ src/com/byronkatz/reap/general/Utility.java
@@ -11,6 +11,7 @@ import android.content.ActivityNotFoundException;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.util.Log;
 import android.view.View;
 import android.view.Window;
 import android.view.WindowManager;
@@ -221,9 +222,13 @@ public class Utility {
   }
 
   public static String displayPercentage(Double value) {
+    
+    //Seems that numberFormat has a beef with Double values.  Downgrade to Float so it doesn't choke
+    
     percentFormat = NumberFormat.getPercentInstance(Locale.US);
     percentFormat.setMaximumFractionDigits(4);
-    result = percentFormat.format(value);
+    result = percentFormat.format(value.floatValue());
+//    Log.d("Utilityclass", "result: " + result + " value: " + value);
     return result;
   }
